# docs/swagger.yaml
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  dto.Product:
    properties:
      amount:
        type: integer
      color:
        type: string
      createdAt:
        type: string
      description:
        type: string
      imageURL:
        type: string
      price:
        type: number
      productID:
        type: string
      productName:
        type: string
      sellerID:
        type: string
      tag:
        items:
          type: string
        type: array
    type: object
...
  dto.Seller:
    properties:
      address:
        type: string
      city:
        type: string
      name:
        type: string
      payment:
        type: string
      phoneNumber:
        type: string
      province:
        type: string
      score:
        type: number
      sellerID:
        type: string
      surname:
        type: string
      transaction:
        items:
          $ref: '#/definitions/dto.Transaction'
        type: array
      username:
        type: string
      zip:
        type: string
    type: object
  dto.SellerRegisterRequest:
    properties:
      address:
        type: string
      city:
        type: string
      name:
        type: string
      password:
        type: string
      payment:
        type: string
      phoneNumber:
        type: string
      province:
        type: string
      score:
        type: number
      surname:
        type: string
      username:
        type: string
      zip:
        type: string
    type: object
  dto.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
  dto.Transaction:
    properties:
      amount:
        minimum: 0
        type: number
      data:
        type: string
      paymentMethod:
        type: string
      product:
        items:
          $ref: '#/definitions/dto.Product'
        type: array
    type: object
...
  model.Seller:
    properties:
      address:
        type: string
      city:
        type: string
      name:
        type: string
      password:
        type: string
      payment:
        type: string
      phoneNumber:
        type: string
      province:
        type: string
      score:
        type: number
      sellerID:
        type: string
      surname:
        type: string
      transaction:
        items:
          $ref: '#/definitions/model.Transaction'
        type: array
      username:
        type: string
      zip:
        type: string
    type: object
  model.Transaction:
    properties:
      amount:
        minimum: 0
        type: number
      data:
        type: string
      paymentMethod:
        type: string
      product:
        items:
          $ref: '#/definitions/dto.Product'
        type: array
    type: object
info:
  contact: {}
paths:
...
  /seller/:
    get:
      consumes:
      - application/json
      description: Retrieves all sellers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Seller'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all sellers
      tags:
      - seller
    post:
      consumes:
      - application/json
      description: Creates a new seller in the database
      parameters:
      - description: Seller to create
        in: body
        name: seller
        required: true
        schema:
          $ref: '#/definitions/dto.SellerRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Seller'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new seller
      tags:
      - seller
  /seller/{seller_id}:
    get:
      consumes:
      - application/json
      description: Retrieves a seller's data by their ID
      parameters:
      - description: Seller ID
        in: path
        name: seller_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Seller'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a seller by ID
      tags:
      - seller
    put:
      consumes:
      - application/json
      description: Updates an existing seller's data by their ID
      parameters:
      - description: Seller ID
        in: path
        name: seller_id
        required: true
        type: string
      - description: Seller data to update
        in: body
        name: seller
        required: true
        schema:
          $ref: '#/definitions/model.Seller'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Seller'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a seller by ID
      tags:
      - seller
  /seller/{seller_id}/balance:
    get:
      consumes:
      - application/json
      description: Retrieves a seller's total balance by their ID
      parameters:
      - description: Seller ID
        in: path
        name: seller_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  type: number
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a seller's total balance by ID
      tags:
      - seller
  /seller/{seller_id}/transaction:
    post:
      consumes:
      - application/json
      description: Append transaction to seller transactions
      parameters:
      - description: Seller ID
        in: path
        name: seller_id
        required: true
        type: string
      - description: Transaction to append
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/dto.Transaction'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Transaction'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Add a transaction by sellerID
      tags:
      - seller
swagger: "2.0"